refactoring: move nb metadata test to new file
